+++ test/lcd_hd44780/lcd_hd44780_test_common.h
@@ -2,7 +2,7 @@
  * @Author: lukasz.niewelt 
  * @Date: 2023-12-07 16:59:53 
  * @Last Modified by: lukasz.niewelt
- * @Last Modified time: 2023-12-07 21:20:56
+ * @Last Modified time: 2023-12-07 23:24:56
  */
 #include <stdint.h>
 #include "mock_LCD_IO_driver.h"
@@ -19,7 +19,7 @@ extern uint16_t expected_buf_lenght;
 void clear_expected_LCD_Port_delay_dump_data(void);
 uint8_t read_prev_LCD_PORT_state (void);
 uint16_t define_expected_sequence_for_first_15_ms_delay(void);
-uint8_t define_expected_sequence_for_read_write_4_bit_data(uint8_t log_no, uint8_t R_W_data, uint16_t delay);
-uint8_t define_expected_sequence_for_send_cmd_to_LCD(uint8_t log_no, uint8_t cmd,uint8_t expected_readed_data, uint16_t additional_cmd_delay);
-uint8_t define_expected_sequence_for_send_data_to_LCD(uint8_t log_no, uint8_t data,uint8_t expected_readed_data, uint16_t additional_cmd_delay);
-uint8_t define_expect_sequence_for_lcd_def_char(uint8_t log_no,enum LCD_CGRAM CGRAM_char_index,const uint8_t *def_char);
+uint16_t define_expected_sequence_for_read_write_4_bit_data(uint16_t log_no, uint8_t R_W_data, uint16_t delay);
+uint16_t define_expected_sequence_for_send_cmd_to_LCD(uint16_t log_no, uint8_t cmd,uint8_t expected_readed_data, uint16_t additional_cmd_delay);
+uint16_t define_expected_sequence_for_send_data_to_LCD(uint16_t log_no, uint8_t data,uint8_t expected_readed_data, uint16_t additional_cmd_delay);
+uint16_t define_expect_sequence_for_lcd_def_char(uint16_t log_no,enum LCD_CGRAM CGRAM_char_index,const uint8_t *def_char);
